requirements.txt
dnspython==2.7.0
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pymongo==4.10.1
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tomli==2.0.2
tzdata==2024.2
